refactor: 커넥션 가져오는 역할 JdbcTemplateUtil로 분리
