+++ src/main/java/chess/dao/GameDaoImpl.java
@@ -9,14 +9,11 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 
 public class GameDaoImpl implements GameDao {
-    private static final String URL = "jdbc:mysql://localhost:3306/chess";
-    private static final String USER = "user";
-    private static final String PASSWORD = "password";
 
     public TurnDto getTurn() {
         String sql = "SELECT game.turn FROM game WHERE id = '1'";
         String turn = "";
-        try (Connection connection = getConnection();
+        try (Connection connection = JdbcTemplateUtil.getConnection();
              PreparedStatement statement = connection.prepareStatement(sql);
              ResultSet resultSet = statement.executeQuery()) {
 
@@ -34,20 +31,10 @@ public class GameDaoImpl implements GameDao {
         }
     }
 
-    public Connection getConnection() {
-        Connection connection = null;
-        try {
-            connection = DriverManager.getConnection(URL, USER, PASSWORD);
-        } catch (SQLException e) {
-            e.printStackTrace();
-        }
-        return connection;
-    }
-
     public void changeTurn() {
         String changeTurnSql = "UPDATE game SET turn = ? WHERE id = ?";
 
-        try (Connection connection = getConnection();
+        try (Connection connection = JdbcTemplateUtil.getConnection();
              PreparedStatement statement = connection.prepareStatement(changeTurnSql)) {
             TurnDto currentTurn = getTurn();
             statement.setString(1, PieceColor.getOpposite(currentTurn.getTurn()));
