+++ src/test/java/chess/dao/GameDaoTest.java
@@ -3,17 +3,9 @@ package chess.dao;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import chess.dto.TurnDto;
-import java.sql.Connection;
 import org.junit.jupiter.api.Test;
 
 class GameDaoTest {
-
-    @Test
-    void connection() {
-        GameDaoImpl gameDao = new GameDaoImpl();
-        Connection connection = gameDao.getConnection();
-        assertThat(connection).isNotNull();
-    }
 
     @Test
     void getTurn() {
